bugfix and new feature ;)
if the auth is enabled, the streamproxy cant access the webif, because he has no password. And because the Streamproxy only request pages on a lokal iface, now all requests on 127.0.0.1:<port> will not require a password.

If the auth is enabled, all extrenal iface still supports the auth, depending on what is configured.