svdrp: initial import from my external git repository
basic remote control interface using the simple vdr protocol, not
feature complete nor very readable, but more a proof of concept.
basic commands however are supposed to remain fully functional and (as
long as enigma2 core supports it) be added as needed.
due to the still experimental nature of this plugin it is currently not
included in the build process.