pluginsort: support resorting eventinfo/movielist/extensions
accessible via menu key in pluginbrowser, if movieepg is installed it
also needs to be updated for resorting the extensions list to work.

also bumps required e2 version to one with weight support because the
additional code overhead was just too much. just upgrade, it's not evil
or anything :D