mphelp: initial checkin
this is a wrapper around the mphelp screen as introduced in the
autotimer. allows plugins to (more or less) dynamically register help
screens and display them.
possible future extensions are a help overview (display all the
registered help screens and allow direct access) and help on e2 core
functionality (possibly driven by an online source for easier
maintencance).