mphelp: allow to use xml as container for the help
instantiate reader using XMLHelpReader(filename). obj[0] is
getNameFunc, obj[1] is getPagesFunc, obj[3] is optional skin name. so
doing registerHelp(*XMLHelpReader(filename)) is possible, with the full
configuration being inside the xml.
be aware that this might raise an exception on parsing errors.

a sample xml help will follow in a subsequent commit. the format was
chosen to be similar to existing enigma2 xml formats (e.g. the wizard)
and to be compatible with xml2po.py without any changes.