Toplevel.py: fix for newer versions of twisted-web
POST is used for those requests to circumvent the cache, but with
twisted-web 10.0.0 and newer only GET and HEAD are valid request
methods. So if we detect a render_GET implementation we add our own
render_POST which in turn just calls the render_GET implementation.
Otherwise we just use the plain static.File implementation as was
done before.