ecasa: add initial support for flickr
requires python-flickrapi which pulls a lot of dependencies and you
currently (as the situation reg. sharing an api key is unknown to me)
have to enter an api key on your own. you can request one from the
flickr homepage.
also adds the previously missing dependency on python-twisted-web which
is present in most images anyway, but better safe than sorry :P