Improve the changed message string.
There is no need to add whitespaces if you add a newline, in the best
case its useless, in the worst it adds two newlines which was most
likely not the intention of the change.
This amends commit 0e1c77ae190605a6c45cf5c0bacb2c6c2de1e8bf.