SwitchService.py: evil voodoo for playing recordings in movie player
seriously, this code is evil - don't copy it. better yet, don't even
look at it. :D

if we assume the sref to be a recorded movie, check if the currently
active screen is the infobar. if it is, open a new movie player instance
and have it start playback. otherwise just let the navigation take care
of things like before.
if this is no recording and the movie player is the current screen,
close it - either way: start playback.

AGAIN: DO NOT COPY. BAD MOJO!