toolkit: add NotifiablePluginBrowser
since the autotimer now uses a notification its behavior is unexpected
for existing users - add a simple extension of the plugin browser to the
toolkit which allows it to display notifications. it's not part of the
autotimer since it might be useful to other plugins and the autotimer
depends on the toolkit anyway.