reconstructapsc: Use enigma2's internal reconstruction, (thanks Milo)
The re-index function is now integrated in E2, preventing a lot of code
duplication. Just use a normal "task" to perform it.

The plugin now lacks the function to do this for whole directories, if
you want that back, just ask and a bit of Python code will add that back.