[autotimer] Less debug output.
On systems with many autotimers / timers, the amount of debug output
going to the serial console ports can negatively affect the performance.

For example, on a system with about 50 autotimers, generating somewhere
between 180 and 220 timers per week, the amount of serial debug output
can result in the spinner being displayed for several seconds.

Signed-off-by: Peter Urbanec <git.user@urbanec.net>