Series2Folder v1.2 - single-action plugin descriptors
Add new PluginDescriptors for single-action "move series to folder" and "move
selected series to folder".

Split out the "move series" action code into a separate class to allow
for the actions, and as a first step to allow automation.

Allow the "move series to folder" and "move selected series to folder" actions
to be configured in or out.