[MountManager] Handle newline in /etc/hostname correctly
In Change hostname, the current code reads the whole /etc/hostname
as the text for the VirtualKeyBoard and doesn't strip the newline(s).

That means that there's a "hidden" newline in the VirtualKeyBoard
text that has to be skipped back over before the text can be edited.

This commit reads only the first line of /etc/hostname
(as does the Busybox hostname -F /etc/hostname) and strips the
newline from it before passing it to VirtualKeyBoard.

The newline is added back when /etc/hostname is written.