EPG Search: Add optional orbital position to search list
[EPGSearch]

Add code to allow the optional display (default enabled) of satellite
orbital positions.

Column space for the orbital position is taken from the left end
of the event title.

Orbital display is disabled for (non-satellite) Beyonwiz images.

[__init__]

Add config variable for the display of orbital position, and
allowShowOrbital to flag whether display of orbital position is
allowed in the image.

[EPGSearchSetup]

Add setup entry for display of orbital position, conditional on
__init__.allowShowOrbital.