[epgsearch] Speed up searches with large numbers of services
For searches with a small nymber of services in the filter, use the
previous search algorithm of extracting each service's EPG data and
doing the event name match in python.

For services with a larger number of services in the filter, use
the low-leven EPG search by name and filter out the unwanted results.

Display a "Searching.." text item in the list while the search is
running to give feedback to the user when a search taks a long time
to run.