Fix crash when Audio Sync settings are selected on a HDMI IN service
Problem found on Beyonwiz T4 with HDMI IN.

When HDMI IN is the service being viewed live, enter
the "Select audio track" popup (long-AUDIO
on Beyonwiz image, my be short-AUDIO in other images).

Press BLUE Audio Sync, crash.

The problem is due to eServiceHDMI::audioTracks()
returning None in Python. This condition is not correcty tested for
in AC3delay.getAudioInformation() and the UI crashes
when oAudioTracks.getCurrentTrack() is called when
oAudioTracks is None.

The problem is also likely to occur when Xine services are
being played, since eServiceXine::audioTracks() also returns None.

The fix ensures that oAudioTracks.getCurrentTrack() is only
called when oAudioTracks is not None, and that other values
normally set in AC3delay.getAudioInformation() are given
sensible default values when oAudioTracks is None and
other iPlayableService audio-related methods
return None in Python.