Fix screen-scraping character encoding
Fix mangling of multi-byte utf-8 encoded characters.

Add a test for the "charset=" <meta/> information so that it works
for both HTML 4 and HTML 5. Previous code only worked for HTML 4.

Use "in" for the search instead of a regular expression, because
the match string doesn't contain RE operators.

When processing a page that's already in utf-8, encode the translation
of named HTML entities into utf-8 instead of trying to put unicode
characters into the 8-bit encoded HTML string.

If the page is already utf-8 encoded, don't re-encode it into utf-8.