Fix screen-scraping of general info
Add utf-8 decodings of some HTML entities to allow them to be edited
out of the text.

Update the extra info regular expressions and the code that cleans
up the matched data for both display and text file output.

Split out the awards regular expression, because it needs to be
handled differently from the other fields.

In the scraped data affected by these changes, clean up spacing
using
' '.join(<string>.split()
rather than the stripmask regular expression. It has the advantage
of also doing the equivalent to a <string>.strip() as well.

Add additional spacing to the extrainfo display to improve readability.