Make IMDB "Select from EPG" function
In the IMDB plugin, MENU>Select from EPG doesn't appear to do
anything useful.

These changes allow its apparently intended function to work.

That function appears to be to allow a "similars" lookup on the
current event on a service to be made, and then GREEN Lookup to do
an IMDb search on the event in focus.

This doesn't appear to be all that useful to me, but at least it
now does something.

When a service is selected, fetch its event id to pass as a parameter
to the "similars" lookup.

Add eventid as an optional parameter to the creation of IMDBEPGSelection,
and pass it as a paramter to the creation of the superclass. Also
pass the serviceref as a string on the creation of the superclass,
since the string form of the serviceref is required for a "similars"
EPGSelection.

Override greenButtonPressed() as well as addTimer(), since the
latter doesn't appear to be called on GREEN in either the OpenViX
or Beyonwiz firmware.

Retain addTimer() as a call of greenButtonPressed() for backwards
compatibility.

Change the null body of onSelectionChanged() to call onSelectionChanged()
in the superclass, and then over-ride the GREEN button hint label
with "Lookup".