Allow IMDb search from MovieSelection
Optionally (default enabled) allow searches of the current
item from the MovieSelection menu or from an assigned coloured
or other configurable button.

Change pluginlist into a tuple of tuples containing
(descriptorEnablerConfig, pluginDescriptor) and use that in both
Plugins() and IMDbSetup.keySave() to determine whether the plugin
is displayed or not.

When a search is initiated from MovieSelection and
the "event name" ends in an extension in MovieList.KNOWN_EXTENSIONS,
remove the extension, and replace any non-alphanumerics in the
event name with spaces.

Tidy up spaces in search string after tags have been removed.