[epgsearch] Make EPG Search - Import from EPG work
In the EPG Search plugin, MENU>Import from EPG work doesn't appear to do
anything useful.

These changes allow its apparently intended function to work.

That function appears to be to allow a "similars" lookup on the
current event on a service to be made, and then GREEN Lookup to do
an IMDb search on the event in focus.

This doesn't appear to be all that useful to me, but at least it
now does something.

When a service is selected, fetch its event id to pass as a parameter
to the "similars" lookup.

Pass the string form of the ref parameter to EPGSelection, since
that's what's required of the "similars" form of EPGSelection.

Override greenButtonPressed() as well as addTimer(), since the
latter doesn't appear to be called on GREEN in either the OpenViX
or Beyonwiz firmware.

Retain addTimer() as a call of greenButtonPressed() for backwards
compatibility.

Add method onSelectionChanged() to call onSelectionChanged()
in the superclass, and then over-ride the GREEN button hint label
with "Search".

Same bug and fix as recently applied to IMDb search.