Avoid over-writing existing files in destination directory
Change the code that generates the list of files to move to allow
a file to be changed from name.ext to name_NNN.ext to avoid overwriting
simlarly named files already in the destination directory.

Use symbolic names for the ".ts" and ".meta" extensions.

Make named frozensets for the different classes of recording filename
extensions.

Change version number to 1.8.