[Networkbrowser] Cleanup of network address handling
Put common code to generate the string form of the network address
in x.x.x.x/netmaskWidth format into a new method makeIpStr().

Check whether both the returned IP address and netmask for error
returns from iNetwork.getAdapterAttribute().

If a network attribute can't be found by iNetwork.getAdapterAttribute()
it returns None, not an empty list or tuple. Change the tests for
the error return from

if len(retval):

to simply

if retval:

so that either a None or an empty return value will fail the test
correctly, and add tests on the length of the IP address
and netmask to be equal to 4.

Simplify the generation of the netmask bitcount by removing the
unnecesary call of formatIp().

Simplify the generation of the network IP address string by using
a list comprehension instead of enumerating all the IP address
and netmask parts.

Simplify getNetworkIPs() by simply returning the return
value of netscan.netzInfo() rather than appending. Also
avoid a potential crash if makeIpStr() returns None.