[Networkbrowser] Prevent near-infinite loop & crash
Check whether iNetwork.getAdapterAttribute() returns an all-zero
IP address or an all-zero netmask, and give an error popup in either
case.

This prevents nmap from enumerating the whole IPv4 address space
if the network scan is run when there is no network interface
available.

If this happens, nmap also returns one line of XML data for each
IP address scanned, which is accumulated in the Console containing
the run of nmap. This eventually leads to memory being completely
consumed and a crash of the UI.