Fix issue #716: [NetworkBrowser] Improvements to default local sharenames
Construct local sharenames for both NFS and CIFS shares in the form
hostname_sharename. The sharename for NFS shares is the basename
of the exported path, or "root" if the exported path is "/".

The code previously used sharename@hostname for CIFS shares and
hostname for NFS shares. The "@" in the CIFS local sharename isn't
a legal character in local sharenames and is removed by MountEdit.