Fix bug #717: [NetworkBrowser] Having disabled automounts can stop the automounter
In AutoMount.CheckMountPoint(), change the logic of updating the
automounter from using /etc/init.d/autofs star & stop to use
/etc/init.d/autofs reload once if any automounts are present.

Split CheckMountPoint() into CheckMountPoint(), which remains the
API call point and initialises self.autofsreload, and doCheckMountPoint()
that carries out the function previously performed by CheckMountPoint().
The attribute self.autofsreload flags when an autofs reload is
needed.

Simplify the tests that construct the command list in doCheckMountPoint()
so that there's no attempt to predict whether the command list will
be instructed. The command list is simply constructed as required
and then tested to check whether it remains empty.

Change the test

if command is not None:

to

if command:

because command is initialised to [], not None. This avoids unnecessary
sleep commands in the command list.