Fix screenscraper RE
The order of some fields in the IMDb screen has changed.

Change the order of the matches in the extrainfomask regular
expression (RE).

The moved parts of the RE are unchanged, just their position in the
RE has changed.

Fixes the loss of some fields (e.g. Trivia, User Reviews) in
the Extra Info screen.