[EPGRefresh] Make ActionMaps compatible with new ConfigListScreen
Changes to ActionMaps in recent changes to ViX ConfigListScreen
caused problems in ViX 5.4.001.002.

An ActionMap in EPGRefreshServiceEditor updated to -2, so that its
actions override the corresponding actions in ConfigListScreen.
Otherwise the save and cancel actions in ConfigListScreen are used
insted of those in EPGRefreshServiceEditor.

Added closeRecursive() methods to EPGRefreshConfiguration and
EPGRefreshServiceEditor to override those methods on the cancel
action in versions of ConfigListScreen that have them.

Backwards compatibility tested in ViX 5.4.000.004, which predates
the ConfigListScreen changes.