Make file screen handling more flexible
Add a new FileScreens module to describe how Series2Folder interacts
with screens that display parts fo the filesystem.

Use FileScreens.activeFileScreens() to avoid background runs when
file screens are in use, and call the appropriate function for the
movie selection screen to update the screen after a manual run of
Series2Folder has moved items. Search the dialog stack and current
session dialog to look for file screens instead of assuming that
the screen can only be the current dialog, and remove hard-coded
method calls on MovieSelection.

Add *args to Series2Folder manual run functions to ignore any
additional named parameters: this is required for EnhancedMovieCenter,
which adds an additional selections parameter to WHERE_MOVIELIST
plugin calls.

Extend the MsgBox method to allow the severity parameter to be
passed through it (used for a warning popup if no suitable file
screen update method can be found.

Fix up comment in __init__.py.

Bump version number to 1.11dev1.