Add support for multi-file selection in Beyonwiz MovieSelection
Allow for portability across PluginDescriptors that have the "multi" kwarg
(like Beyonwiz) and those that don't.

Allow for action to optionally carry a list of servicerefs that contains
the currently selected items in the media selection list, if there
are selections.

Extend Series2FolderActionsBase.moveSelection to be a set of
servicerefs, rather than either a serviceref or None.

Bump version number to 1.11dev2.