Default to a single Series to Folder... plugin entry in the MovieSelection menu
[__init__]

Change the defaults for config.plugins.seriestofolder.showmovebutton
and showselmovebutton to false, so that they no longer appear in
the button selectors or the MovieSelection menu.

[plugin]

Clarify the description text for
config.plugins.seriestofolder.showmovebutton and showselmovebutton
in Series2FolderConfig it include what the menu item and button
names for the menu/button actions will be.

[plugin, control]

Bump version number to 1.11dev6.