Add message about how to enable background mode
[__init__]

Add config variable config.plugins.seriestofolder.autoreminder
defaulted to 5 to count down the number of times the message has
been displayed.

Add a config.plugins.seriestofolder.auto to set the countdown to 0
when background mode is enabled.

[plugin]

Add a message to Series2Folder to tell the user how to enable
background mode. The message is shown a limited number of times,
controlled by config.plugins.seriestofolder.autoreminder, and will
also not be shown after background has been enabled (even if it
later disabled).

Background mode is not enabled by default, because there is no
obvious opportunity for the user to disable it before it first runs.

[plugin, control]

Bump version number to 1.11dev9.